+++ packages/vm/src/lib.rs
@@ -4,7 +4,7 @@ mod calls;
 mod compatability;
 mod context;
 mod conversion;
-pub mod errors;
+mod errors;
 mod instance;
 mod memory;
 mod middleware;
@@ -17,6 +17,7 @@ pub use crate::cache::CosmCache;
 pub use crate::calls::{
     call_handle, call_handle_raw, call_init, call_init_raw, call_query, call_query_raw,
 };
+pub use crate::errors::{VmError, VmResult};
 pub use crate::instance::Instance;
 pub use crate::modules::FileSystemCache;
 pub use crate::serde::{from_slice, to_vec};